Error embed instead of pointing to the Github, point to the repo variable/issues link.
When linking through steamid now checks if it can be converted to integer and replies to the user complaining about bot no likey like worderinos.

Changed a bit the error embed for linking through vanity and steamid.
On steamid added the help command at the title.

On vanity added the help command at the title and also added an example to remind users to not use the whole URL, the contents are a bit spaced so people feel less overwhelmed on reading.